refactor: git service
move all git operations into the git service
