+++ pkg/agent/toolbox/git/add.go
@@ -4,8 +4,8 @@
 package git
 
 import (
+	"github.com/daytonaio/daytona/pkg/git"
 	"github.com/gin-gonic/gin"
-	"github.com/go-git/go-git/v5"
 )
 
 func AddFiles(c *gin.Context) {
@@ -15,32 +15,13 @@ func AddFiles(c *gin.Context) {
 		return
 	}
 
-	repo, err := git.PlainOpen(req.Path)
-	if err != nil {
-		c.AbortWithError(400, err)
-		return
+	gitService := git.Service{
+		ProjectDir: req.Path,
 	}
 
-	w, err := repo.Worktree()
-	if err != nil {
+	if err := gitService.Add(req.Files); err != nil {
 		c.AbortWithError(400, err)
 		return
-	}
-
-	if len(req.Files) == 1 && req.Files[0] == "." {
-		_, err = w.Add(".")
-		if err != nil {
-			c.AbortWithError(400, err)
-			return
-		}
-	} else {
-		for _, file := range req.Files {
-			_, err = w.Add(file)
-			if err != nil {
-				c.AbortWithError(400, err)
-				return
-			}
-		}
 	}
 
 	c.Status(200)
