+++ pkg/agent/toolbox/git/pull.go
@@ -4,8 +4,9 @@
 package git
 
 import (
+	"github.com/daytonaio/daytona/pkg/git"
 	"github.com/gin-gonic/gin"
-	"github.com/go-git/go-git/v5"
+	go_git "github.com/go-git/go-git/v5"
 	"github.com/go-git/go-git/v5/plumbing/transport/http"
 )
 
@@ -16,31 +17,20 @@ func PullChanges(c *gin.Context) {
 		return
 	}
 
-	repo, err := git.PlainOpen(req.Path)
-	if err != nil {
-		c.AbortWithError(400, err)
-		return
-	}
-
-	w, err := repo.Worktree()
-	if err != nil {
-		c.AbortWithError(400, err)
-		return
-	}
-
-	options := &git.PullOptions{
-		RemoteName: "origin",
-	}
-
+	var auth *http.BasicAuth
 	if req.Username != nil && req.Password != nil {
-		options.Auth = &http.BasicAuth{
+		auth = &http.BasicAuth{
 			Username: *req.Username,
 			Password: *req.Password,
 		}
 	}
 
-	err = w.Pull(options)
-	if err != nil && err != git.NoErrAlreadyUpToDate {
+	gitService := git.Service{
+		ProjectDir: req.Path,
+	}
+
+	err := gitService.Pull(auth)
+	if err != nil && err != go_git.NoErrAlreadyUpToDate {
 		c.AbortWithError(400, err)
 		return
 	}
