+++ pkg/git/pull.go
@@ -1,0 +1,29 @@
+// Copyright 2024 Daytona Platforms Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+package git
+
+import (
+	"github.com/go-git/go-git/v5/plumbing/transport/http"
+
+	"github.com/go-git/go-git/v5"
+)
+
+func (s *Service) Pull(auth *http.BasicAuth) error {
+	repo, err := git.PlainOpen(s.ProjectDir)
+	if err != nil {
+		return err
+	}
+
+	w, err := repo.Worktree()
+	if err != nil {
+		return err
+	}
+
+	options := &git.PullOptions{
+		RemoteName: "origin",
+		Auth:       auth,
+	}
+
+	return w.Pull(options)
+}
